de-duplicate things
