+++ app/lish/[did]/[publication]/PublicationNav.tsx
@@ -1,4 +1,5 @@
 import { SpeedyLink } from "components/SpeedyLink";
+import { sortPublicationPages } from "./sortPublicationPages";
 
 export type PublicationNavPage = {
   id: number;
@@ -13,12 +14,7 @@ export function PublicationNav(props: {
 }) {
   if (props.pages.length === 0) return null;
 
-  let sortedPages = [...props.pages].sort((a, b) => {
-    let aHome = a.path === "/" ? 0 : 1;
-    let bHome = b.path === "/" ? 0 : 1;
-    if (aHome !== bHome) return aHome - bHome;
-    return (a.path ?? "").localeCompare(b.path ?? "");
-  });
+  let sortedPages = sortPublicationPages(props.pages);
 
   return (
     <nav className="publicationNav border-t border-b border-border-light w-full sm:max-w-[calc(var(--page-width-units)*1.25)] mx-auto">
