+++ app/lish/[did]/[publication]/page.tsx
@@ -9,8 +9,7 @@ import React from "react";
 import { NotFoundLayout } from "components/PageLayouts/NotFoundLayout";
 import { normalizePublicationRecord } from "src/utils/normalizeRecords";
 import { PublicationContent } from "./PublicationContent";
-import { PublicationPageRenderer } from "./[rkey]/PublicationPageRenderer";
-import { PubLeafletPublicationPage } from "lexicons/api";
+import { tryRenderPublicationPage } from "./tryRenderPublicationPage";
 import {
   PublicationThemeProvider,
   PublicationBackgroundProvider,
@@ -54,27 +53,13 @@ export default async function Publication(props: {
 
   if (!publication) return <PubNotFound />;
 
-  let homePage = publication.publication_pages?.find(
-    (p) => p.path === "/" && p.record_uri && p.record,
-  );
-
   try {
-    if (homePage && homePage.record) {
-      let pageRecord =
-        homePage.record as unknown as PubLeafletPublicationPage.Record;
-      return (
-        <PublicationPageRenderer
-          did={did}
-          page={{
-            id: homePage.id,
-            path: "/",
-            title: homePage.title,
-            record: pageRecord,
-          }}
-          publication={publication}
-        />
-      );
-    }
+    const homePageRender = tryRenderPublicationPage({
+      did,
+      publication,
+      path: "/",
+    });
+    if (homePageRender) return homePageRender;
     return (
       <PublicationThemeProvider
         theme={record?.theme}
